Refactored to integrate upload_to_cloud.py into base_loader.py
